fix 64bit error
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@29728 72102866-910b-0410-8b05-ffd578937521

+++ Source/Additions/GCDictionary.m
@@ -134,7 +134,7 @@ _GCDescribeObjects(NSMapTable *table, const GCInfo *objectStruct)
 }
 
 static const NSMapTableKeyCallBacks GCInfoMapKeyCallBacks = {
-  (unsigned(*)(NSMapTable *, const void *))_GCHashObject,
+  (NSUInteger(*)(NSMapTable *, const void *))_GCHashObject,
   (BOOL(*)(NSMapTable *, const void *, const void *))_GCCompareObjects,
   (void (*)(NSMapTable *, const void *))_GCRetainObjects,
   (void (*)(NSMapTable *, void *))_GCReleaseObjects,
